Fixed scaling of 2.1 nsx signals, and added 'scaling' option to read_*
File spec 2.1 can be converted from digital sample to voltage using the
digitization factor attribute of nev files. If no nev file exists, only
raw samples can be retourned. For this reason the option
scaling={'raw'|'voltage'} was added, so that even in the abesnce of a
nev file, nsX data can be read as raw sample values.